add README

+++ src/main/java/com/comp2024b/tocountornot/bean/Card.java
@@ -11,5 +11,5 @@ public class Card {
     private double income;
     private double expense;
     private double balance;
-    private Long uid;
+    private int uid;
 }
